It seems cPickle is not a valid module in jupyter
I just heard about this project on Linear digressions and when I came home to try it out it won't load and raises this: 
----> 1 import cPickle as pickle
         2 import os
         3 import sys
         4 import errno
         5 

ModuleNotFoundError: No module named 'cPickle'

I propose changing it to _pickle. 
I'm using python 3.6. I also am not experienced at this, it is just that I want it to work (I don't even know how to test it the way I proposed).