Syntax errors
I managed to discover how to test the changes and after I changed cPickle to _pickle it raised some syntax errors in this file. Just some parentheses missing. I don't know if it was supposed to be like that but now it ran without raising anything.

Thanks for this cool project!